// src/features/wallet/models/account/types/AccountLoadState.ts
export enum AccountLoadState {
  IDLE = 'IDLE',
  ACCOUNT_REQUESTED = 'ACCOUNT_REQUESTED',
  ACCOUNT_DETECTION_FAILED = 'ACCOUNT_DETECTION_FAILED',
  LOCKED = 'LOCKED',
  UNLOCK_REQUESTED = 'UNLOCK_REQUESTED',
  UNLOCK_REJECTED = 'UNLOCK_REJECTED',
  WAITING__UNLOCK = 'WAITING_UNLOCK',
  UNLOCK_FAILED = 'UNLOCK_FAILED',
  ACCOUNT_LOADED = 'ACCOUNT_LOADED',
}